+++ wayland-scanner/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Bump bitflags to 2.0
 - Remove `io-lifetimes` from the generated code following wayland-backend 0.3 dropping it.
+- Generate `BorrowedFd<'_>` arguments instead of `RawFd`
 
 # 0.30.1 -- 2023-06-17